Show logout as a slot
